Rename to Logger, add log()
